feat : add supplier update and delete

+++ src/Components/deleteHandler.js
@@ -1,8 +1,25 @@
 import api from "../Util/api";
 
 export const deleteHandler = (id, reloader, url) => {
-  api().delete(`api/${url}/${id}`).then(response => {
-    window['toastr'].success(response.data.message);
-    reloader(prev => ({...prev}));
-  });
+  window['Swal'].fire({
+    title: 'Are you sure?',
+    text: "You won't be able to revert this!",
+    icon: 'warning',
+    showCancelButton: true,
+    confirmButtonColor: '#3085d6',
+    cancelButtonColor: '#d33',
+    confirmButtonText: 'Yes, delete it!'
+  }).then((result) => {
+    if (result.isConfirmed) {
+      api().delete(`api/${url}/${id}`).then(response => {
+        window['toastr'].success(response.data.message);
+        reloader(prev => ({...prev}));
+      });
+      window['Swal'].fire(
+        'Deleted!',
+        'Your file has been deleted.',
+        'success'
+      )
+    }
+  })
 }
